moved ComponentSet from Screen.js to Components.js
